requirements.txt
pip==1.5.6
setuptools==7.0
django==1.8.5
requests==2.8.1
djangorestframework==3.3.0
django-ipware==1.1.1
user-agents==1.0.1
#mysql-python==1.2.5
django-rest-swagger==0.3.4